requirements.txt
Django
djangorestframework
cherrypy
